added a search_in_stack to replace is_locked and search_node, removed the file locked_bonus.c
