+++ bonus/srcs_bonus/main_bonus/cub3d_bonus.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 10:36:55 by mhotting          #+#    #+#             */
-/*   Updated: 2024/06/27 12:43:23 by brappo           ###   ########.fr       */
+/*   Updated: 2024/06/27 13:34:19 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
