+++ bonus/srcs_bonus/pathfinding_bonus/pathfinding_bonus.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 19:09:13 by brappo            #+#    #+#             */
-/*   Updated: 2024/06/27 12:43:00 by brappo           ###   ########.fr       */
+/*   Updated: 2024/06/27 13:36:30 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,16 @@
 #include "libft.h"
 
 #include <unistd.h>
+
+void	lock_tile(t_pathfinding *pathfinding)
+{
+	t_stack_path	*new_stack;
+
+	new_stack = pathfinding->stack->next;
+	pathfinding->stack->next = pathfinding->locked_tiles;
+	pathfinding->locked_tiles = pathfinding->stack;
+	pathfinding->stack = new_stack;
+}
 
 t_list	*find_path(t_mlx_coords *start, t_mlx_coords *end, t_map *map)
 {
