+++ bonus/srcs_bonus/pathfinding_bonus/t_pathfinding_utils_bonus.c
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 11:59:16 by brappo            #+#    #+#             */
-/*   Updated: 2024/06/26 15:38:35 by brappo           ###   ########.fr       */
+/*   Updated: 2024/06/27 13:37:37 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,4 +40,16 @@ void	t_pathfinding_init(t_pathfinding *pathfinding, t_mlx_coords *start,
 	pathfinding->end = end;
 	pathfinding->start = start;
 	add_path_node(start, pathfinding, NULL, 0);
+}
+
+t_stack_path	*search_in_stack(t_stack_path *stack, t_mlx_coords *position)
+{
+	while (stack)
+	{
+		if (stack->position.x == position->x
+			&& stack->position.y == position->y)
+			return (stack);
+		stack = stack->next;
+	}
+	return (NULL);
 }
